Elaborate on project management and billing
No developer will work in a company where the mentality is "you are dumb; you are responsible for fixing your stupid mistakes; you won't get paid to learn". If a customer wants me to build him a spaceship the best I can do is tell him "hey, I don't know how to build a spaceship; but if you're willing to work with me, we can find out together, one small piece at a time".

This is the promise of Agile and how many successful companies work nowadays. We promise to do our best, but acknowledge that we don't know everything, mistakes happen and requirements change.

Instead of instilling the mentality of "I should know everything and mistakes are terrible", embrace change and failure and learn from them.